I SWEAR this isn't spaghetti code.
Finished the menu. Sorry it doesn't look the best, I could probably work a little harder to do that haha! I did add a "high score" feature though.

Next plan is to add a pause menu during the game.

(It may or may not lag a little because of my spaghetti code ^_^ )